BLE update 0.3

+++ T-TWR-master/examples/Factory/Factory.ino
@@ -24,7 +24,7 @@ using namespace ace_button;
 U8G2_SH1106_128X64_NONAME_F_HW_I2C  u8g2(U8G2_R0, U8X8_PIN_NONE);
 
 
-int bitpack[512];
+uint8_t bitpack[512];
 bool bitready = false;
 
 List list;
